[BSR] ProfileForm: split components
  to make it easier to poc on formik form, i needed to extract the
ProfileFrom from the ProfileInformationsModal.
