+++ src/components/pages/Home/ProfileAndSupport/ProfileForm/index.js
@@ -1,0 +1,1 @@
+export { default } from './ProfileForm'
